Move getAidboxToken and LoginService to setupTests
