+++ src/services/fhir.ts
@@ -7,11 +7,8 @@ import {
 } from 'aidbox-react';
 import type { AxiosRequestConfig } from 'axios';
 
-import { User } from '@beda.software/aidbox-types';
-import { ensure, initServicesFromService } from '@beda.software/fhir-react';
-import { RemoteData, RemoteDataResult, Token } from '@beda.software/remote-data';
-
-export type LoginService = (user: User) => Promise<RemoteData<Token>>;
+import { initServicesFromService } from '@beda.software/fhir-react';
+import { RemoteDataResult } from '@beda.software/remote-data';
 
 const fhirService = async <S = any, F = any>(config: AxiosRequestConfig): Promise<RemoteDataResult<S, F>> => {
     return aidboxService({ ...config, baseURL: axiosInstance.defaults.baseURL + '/fhir' });
@@ -51,15 +48,5 @@ export const {
     applyFHIRServices,
     service,
 } = initServicesFromService(fhirService);
-
-export async function getAidboxToken(user: User, loginService: LoginService): Promise<Token> {
-    if (!user.email) {
-        throw new Error('Can not login for user without an email');
-    }
-
-    const result = await loginService(user);
-
-    return ensure(result);
-}
 
 export { aidboxService, axiosInstance, setInstanceToken, resetInstanceToken, setInstanceBaseURL };
